% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{compute_significance_data}
\alias{compute_significance_data}
\title{Compute associations with phenotypical variables.}
\usage{
compute_significance_data(values, pdata, component_names)
}
\arguments{
\item{values}{A matrix containing either principal components or surrogate
variables.}

\item{pdata}{A data.frame containing the phenotypical data for the samples.}

\item{component_names}{A character vector containing the component names.}
}
\value{
A tbl_df with the results of the association tests.
}
\description{
Compute associations with phenotypical variables.
}

