% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{compute_significance_data_var_names}
\alias{compute_significance_data_var_names}
\title{Compute associations with phenotypical variables.}
\usage{
compute_significance_data_var_names(values, pdata, component_names, var_names,
  method = c("lm", "kruskal"))
}
\arguments{
\item{values}{A matrix containing either principal components or surrogate
variables.}

\item{pdata}{A data.frame containing the phenotypical data for the samples.}

\item{component_names}{A character vector containing the component names.}

\item{var_names}{Variable names to be used for calculating the p_values.}

\item{method}{A character indicating the method used for the association
tests.}
}
\value{
A data.frame with the results of the association tests.
}
\description{
Compute associations with phenotypical and control variables in order to
find out which of them are significant, and thus possibly introducing
some confounding in the data.
}
\details{
The current implementation allows the user to use two different methods
for testing association: a general linear model (lm) and a Kruskal-Wallis
test. User should be careful, as the Kruskal Wallis test should not be used
for continuous independent variables.

An association test is performed between every component (Principal Component
or Surrogate Variable) and every phenotype and control variable. P-values are
then collected and stored in a \code{tbl_df} object describing all the
comparisons.

This function receives variable names as a parameter.
}
\keyword{internal}
