% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_analysis.r
\name{cut_pvalues}
\alias{cut_pvalues}
\title{Cut p-values according to a significance criterium.}
\usage{
cut_pvalues(p_values)
}
\arguments{
\item{p_values}{A numeric containing the p-values}
}
\value{
A factor containing the cut categories.
}
\description{
Cut p-values according to a significance criterium.
}

