% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_control_variables}
\alias{get_control_variables}
\title{Get control variables.}
\usage{
get_control_variables(rgset)
}
\arguments{
\item{rgset}{An RGChannelSet object containing the control data.}
}
\value{
A matrix with the values of the control elements.
}
\description{
Get a matrix containing the signals for the control probes of a given
RGChannelSet object.
}
\details{
The function \code{get_control_variables} just takes a \code{RGChannelSet}
object as input, accesses the values of its control probes, and returns
them as a numeric matrix.

This is an internal function. Its purpose is to provide the main analysis
functions with more data to test. It is usually useful to test if control
probes are introducing some confounding in the dataset, as it might be
signaling that there are unknown effects in the data.
}
