% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_analysis.r
\name{get_p_value}
\alias{get_p_value}
\title{Get p-value from f statistic data.}
\usage{
get_p_value(fstat)
}
\arguments{
\item{fstat}{A numeric vector of length three obtained from a linear model
fit.}
}
\value{
The computed p-value.
}
\description{
Get p-value from f statistic data.
}

