% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_analysis.r
\name{get_var_names}
\alias{get_var_names}
\title{Get variable names suitable for analysis.}
\usage{
get_var_names(pdata)
}
\arguments{
\item{pdata}{A data.frame containing the phenotypical data.}
}
\value{
A character vector containing the selected variable names.
}
\description{
Get variable names suitable for analysis.
}

