% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenopart}
\alias{phenopart}
\title{Partial phenotype information for smallmeth}
\format{A data.frame with 25 rows and 6 variables:
\describe{
\item{Sample_Group}{Group of the sample (Normal, Tumor)}
\item{Slide}{Slide in the plate where the sample was placed}
\item{Sex}{Gender of the sample (Male, Female)}
\item{DFS}{Disease Free Survival}
\item{GS}{Glutamine Synthetase}
\item{Age}{Age of the individual}
}}
\usage{
phenopart
}
\description{
A data.frame describing part of the available phenotype information for the
methylation analysis project represented by the smallmeth dataset. Each row
describes a sample, and each column represents a different variable.
}
\keyword{datasets}
