% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_significance}
\alias{plot_significance}
\title{Plot results of SVD/SVA confounder analysis.}
\usage{
plot_significance(x)
}
\arguments{
\item{x}{A list containing the results from the confounder analysis
function.}
}
\value{
A ggplot2 heatmap showing the results of all the association tests.
}
\description{
Plot results of SVD/SVA confounder analysis as a heatmap where each cell
represents the statistical significance of the given association test.
}
\details{
This internal function takes the result of either \code{svd_analysis} or
\code{sva_analysis} functions and returns a heatmap representing the degree
of significance for each of the performed association tests. Components or
surrogate variables are represented on the x axis. Control and phenotype
variables are represented in the y axis. The color of each cell corresponds
to the degree of statistical significance for the corresponding association
test.
}

