% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva_analysis.r
\name{plot_sva_significance}
\alias{plot_sva_significance}
\title{Plot results of SVA confounder analysis.}
\usage{
plot_sva_significance(sva_result)
}
\arguments{
\item{sva_result}{A list containing the results from the sva_analysis function.}
}
\value{
A ggplot2 plot.
}
\description{
Plot results of SVA confounder analysis.
}

