% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_svd_scree}
\alias{plot_svd_scree}
\title{Plot ratio of explained variance.}
\usage{
plot_svd_scree(x)
}
\arguments{
\item{x}{A list containing the results from the svd_analysis
function.}
}
\value{
A ggplot2 scree plot showing the ratio of explained variance for
every Principal Component.
}
\description{
Generates a plot showing the ratio of explained variance by the Principal
Components as a scree plot.
}
\details{
This is an internal function taking the output of the \code{svd_analysis}
function as input. It generates a scree plot showing the ratio of explained
variance for the ordered Principal Components. It is very similar to the
\code{screeplot} function but it uses ggplot2 instead of base graphics.
}
