% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_analysis.r
\name{plot_svd_significance}
\alias{plot_svd_significance}
\title{Plot results of SVD confounder analysis.}
\usage{
plot_svd_significance(svd_result)
}
\arguments{
\item{svd_result}{A list containing the result from the svd_analysis function.}
}
\value{
A ggplot2 plot.
}
\description{
Plot results of SVD confounder analysis.
}

