% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{screeplot.SVAAnalysis}
\alias{screeplot.SVAAnalysis}
\title{Plot ratio of explained variance.}
\usage{
\method{screeplot}{SVAAnalysis}(x, ...)
}
\arguments{
\item{x}{A list containing the results from the sva_analysis function.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generates an error when the user tries to call the \code{screeplot} function
on a SVAAnalysis object.
}
\details{
This is a generic method taking the output of the \code{sva_analysis}
function as input. It generates an error..
}
