% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallmeth}
\alias{smallmeth}
\title{Small methylation example dataset}
\format{A matrix with 3000 rows and 25 columns.}
\usage{
smallmeth
}
\description{
A dataset containing a subset of a DNA Methylation analysis project based on
microarrays. It is used by a vignette in order to show how the package works.
Rows indicate probes and columns samples.
}
\keyword{datasets}
