% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva_analysis.r
\name{sva_analysis}
\alias{sva_analysis}
\title{Perform SVA confounder analysis}
\usage{
sva_analysis(values, pdata, main_formula, null_formula = NULL,
  vfilter = NULL, rgset = NULL)
}
\arguments{
\item{values}{A matrix containing the data to analyze. Rows represent
variables and columns samples.}

\item{pdata}{A data.frame containing the phenotype information for the
samples.}

\item{main_formula}{A formula describing the main model of interest.}

\item{null_formula}{A formula describing the null model of interest.}

\item{vfilter}{The number of most variable rows to use in order to
estimate the number of surrogate variables.}

\item{rgset}{If not NULL, compute association with control probes data
contained in the provided RGChannelSet.}
}
\value{
A list containing the number of Surrogate Variables (num_sv), the
  results of the confounding analysis tests (significance), and the Surrogate
  Variables obtained by SVA (surrogates).
}
\description{
Perform SVA confounder analysis and association tests between the Surrogate
Variables of the given input values and the phenotype data.
}
\details{
This function performs a series of association tests between the Surrogate
Variables of the input values and the provided phenotype data.

If a RGChannelSet is given, it can also compute the association with the
control probes. This implementation uses Linear models.
}
