% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cut_pvalues}
\alias{cut_pvalues}
\title{Cut p-values.}
\usage{
cut_pvalues(p_values)
}
\arguments{
\item{p_values}{A numeric containing the p-values}
}
\value{
A factor containing the resulting categories.
}
\description{
Cut p-values according to a significance criterium.
}
\details{
This is a simple internal function for cutting a vector of p-values into
a set of meaningful categories. The categories are taken from the original
book chapter that inspired the package.
}
\keyword{internal}
