% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_p_value}
\alias{get_p_value}
\title{Get p-value from F statistic data.}
\usage{
get_p_value(fstat)
}
\arguments{
\item{fstat}{A numeric vector of length three obtained from a linear model
fit.}
}
\value{
The computed p-value.
}
\description{
Get p-value from F statistic data in the summary of an \code{lm} object.
}
\details{
This is an internal function that extracts the p-value from the information
generated by a call to the \code{summary} function on a \code{lm} object. It
uses the F statistic value and the degrees of freedom in order to compute the
p-value.
}
\keyword{internal}
