% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_var_names}
\alias{get_var_names}
\title{Get variable names suitable for analysis.}
\usage{
get_var_names(pdata)
}
\arguments{
\item{pdata}{A data.frame containing the phenotypical data.}
}
\value{
A character vector containing the selected variable names.
}
\description{
Get variable names that fulfill the conditions in order to be tested for
association.
}
\details{
This is an internal function taking the phenotype data.frame as input. It
returns a character vector containing the names of those variables which can
be tested for association. The current implementation avoids testing on
categorical variables where all elements belong to the same level, or where
every level has only one observation.
}
\keyword{internal}
