% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_significance}
\alias{plot_significance}
\title{Plot results of confounder analysis.}
\usage{
plot_significance(result)
}
\arguments{
\item{result}{A list containing the results from the confounder analysis function.}
}
\value{
A ggplot2 plot.
}
\description{
Plot results of confounder analysis.
}

