% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_svd_scree}
\alias{plot_svd_scree}
\title{Plot variance explained}
\usage{
plot_svd_scree(svd_result)
}
\arguments{
\item{svd_result}{A list containing the result from the svd_analysis
function.}
}
\value{
A ggplot2 plot.
}
\description{
Plot variance explained
}

