% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sva_analysis.r
\name{sva_analysis}
\alias{sva_analysis}
\title{Perform SVA confounder analysis}
\usage{
sva_analysis(values, pdata, main_formula, vfilter = NULL)
}
\arguments{
\item{values}{A matrix containing the data to analyze. Rows are for variables
and columns for samples.}

\item{pdata}{A data.frame containing the phenotypical information for the
samples.}

\item{main_formula}{A formula describing the main model of interest.}

\item{vfilter}{The number of most variant variables to use in order to
estimate the number of surrogate variables.}
}
\value{
A list containing the number of surrogate variables (num_sv), the
results of the confounding analysis tests (significance), and the surrogate
variables obtained by sva (surrogates).
}
\description{
Perform SVA confounder analysis
}

