% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_analysis.r
\name{svd_analysis}
\alias{svd_analysis}
\title{Perform SVD confounder analysis.}
\usage{
svd_analysis(values, pdata, center = TRUE, scale = FALSE, rgset = NULL,
  method = c("lm", "kruskal"))
}
\arguments{
\item{values}{A matrix of numerical values. Rows represent
variables and columns samples.}

\item{pdata}{A data.frame containing the phenotype data for the samples.}

\item{center}{A logical value indicating whether the variables should be
shifted to be zero centered. Alternately, a vector of length equal the
number of columns of x can be supplied. The value is passed to scale.}

\item{scale}{A logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place. The default
is FALSE for consistency with S, but in general scaling is advisable.
Alternatively, a vector of length equal the number of columns of x can be
supplied. The value is passed to scale. Use when the variables are in
arbitrary units of measurement.}

\item{rgset}{If not NULL, compute association with control probes data
contained in the provided RGChannelSet.}

\item{method}{Method used for computing p-values.}
}
\value{
A list containing the proportion of variance explained by the
  principal components (variance_explained) and a data.frame representing the
  results from the association analysis (significance).
}
\description{
Perform SVD confounder analysis and association tests between the Principal
Components of the given input values and the phenotype data.
}
\details{
This function performs a series of association tests between the Principal
Components of the input values and the provided phenotype data. Previously,
it centers and scales the data, something that might be useful when the
variables are measured in different units.

If a RGChannelSet is given, it can also compute the association with the
control probes. This implementation can use two different methods for testing
(Linear models and Kruskal-Wallis).

The number of meaningful components is computed using the function
\code{EstDimRMT} from the package \code{isva}.
}
